fix: fixed styles for login button
