+++ apps/comps/components/providers.tsx
@@ -7,30 +7,30 @@ import {
   createAuthenticationAdapter,
   darkTheme,
 } from "@rainbow-me/rainbowkit";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import {QueryClient, QueryClientProvider} from "@tanstack/react-query";
 import axios from "axios";
-import { useAtom } from "jotai";
+import {useAtom} from "jotai";
 import React from "react";
-import { type ReactNode, useState } from "react";
-import { createSiweMessage } from "viem/siwe";
-import { WagmiProvider } from "wagmi";
+import {type ReactNode, useState} from "react";
+import {createSiweMessage} from "viem/siwe";
+import {WagmiProvider} from "wagmi";
 
-import { ThemeProvider } from "@recallnet/ui2/components/theme-provider";
+import {ThemeProvider} from "@recallnet/ui2/components/theme-provider";
 
-import { userAtom } from "@/state/atoms";
-import { clientConfig } from "@/wagmi-config";
+import {userAtom} from "@/state/atoms";
+import {clientConfig} from "@/wagmi-config";
 
 const AUTHENTICATION_STATUS: AuthenticationStatus = "unauthenticated";
 const CONFIG = clientConfig();
 
-function WalletProvider(props: { children: ReactNode }) {
+function WalletProvider(props: {children: ReactNode}) {
   const [queryClient] = useState(() => new QueryClient());
-  const [, setUser] = useAtom(userAtom);
+  const [user, setUser] = useAtom(userAtom);
 
   const authAdapter = React.useMemo(() => {
     return createAuthenticationAdapter({
       getNonce: async () => {
-        const { data: res } = await axios<{ nonce: string }>({
+        const {data: res} = await axios<{nonce: string}>({
           baseURL: "",
           method: "get",
           url: "/api/nonce",
@@ -42,7 +42,7 @@ function WalletProvider(props: { children: ReactNode }) {
 
         return res.nonce;
       },
-      createMessage: ({ nonce, address, chainId }) => {
+      createMessage: ({nonce, address, chainId}) => {
         return createSiweMessage({
           domain: document.location.host,
           address,
@@ -53,18 +53,18 @@ function WalletProvider(props: { children: ReactNode }) {
           nonce,
         });
       },
-      verify: async ({ message, signature }) => {
-        const res = await axios<{ ok: boolean; address: string }>({
+      verify: async ({message, signature}) => {
+        const res = await axios<{ok: boolean; address: string}>({
           baseURL: "",
           method: "post",
           url: "/api/login",
           headers: {
             Accept: "application/json",
           },
-          data: { message, signature },
+          data: {message, signature},
         });
 
-        setUser({ address: res.data.address, loggedIn: true });
+        setUser({address: res.data.address, loggedIn: true});
 
         return res.data.ok;
       },
@@ -74,6 +74,7 @@ function WalletProvider(props: { children: ReactNode }) {
       },
     });
   }, [setUser]);
+  console.log('SET FUCKING USER', user)
 
   return (
     <WagmiProvider config={CONFIG}>
@@ -91,7 +92,7 @@ function WalletProvider(props: { children: ReactNode }) {
   );
 }
 
-export function Providers({ children }: { children: React.ReactNode }) {
+export function Providers({children}: {children: React.ReactNode}) {
   return (
     <ThemeProvider
       attribute="class"
