+++ apps/comps/components/siwe/index.tsx
@@ -1,12 +1,70 @@
 "use client";
 
-import { ConnectButton } from "@rainbow-me/rainbowkit";
+import {ConnectButton} from "@rainbow-me/rainbowkit";
 import React from "react";
 
-import { Button } from "@recallnet/ui2/components/shadcn/button";
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from "@recallnet/ui2/components/shadcn/dropdown-menu";
+import {Identicon} from "../Identicon";
+import {Button} from "@recallnet/ui2/components/shadcn/button";
+import {userAtom} from "@/state/atoms";
+import {useAtom} from "jotai";
 
 export const SIWEButton: React.FunctionComponent<
   React.ComponentProps<typeof Button>
 > = (props) => {
-  return <ConnectButton {...props}>{props.children}</ConnectButton>;
+  const [user, setUser] = useAtom(userAtom)
+
+  const handleLogout = () => {
+    setUser({loggedIn: false, address: ""});
+  };
+
+  return <ConnectButton.Custom>{
+    ({openConnectModal}) => {
+      return user.loggedIn ? (
+        <div className="mx-3 flex items-center space-x-3">
+          <span className="text-sm font-medium text-gray-400">STAKED</span>
+          <span className="text-sm font-medium text-white">700</span>
+          <span className="text-sm font-medium text-gray-400">WALLET</span>
+          <span className="text-sm font-medium text-white">1000</span>
+          <Button className="bg-black p-0 text-sky-700 hover:text-sky-600">
+            ADD FUNDS
+          </Button>
+          <DropdownMenu>
+            <DropdownMenuTrigger asChild>
+              <div className="ml-5 flex cursor-pointer items-center justify-between">
+                <Identicon address={user.address} />
+                <div className="focus ml-3 text-sm font-medium text-white">
+                  {user.address.slice(0, 6)}...{user.address.slice(-4)}
+                </div>
+              </div>
+            </DropdownMenuTrigger>
+            <DropdownMenuContent className="bg-gray-900">
+              <DropdownMenuItem
+                onClick={handleLogout}
+                className="cursor-pointer hover:bg-gray-800"
+              >
+                Logout
+              </DropdownMenuItem>
+            </DropdownMenuContent>
+          </DropdownMenu>
+        </div>
+      ) : (
+        <div className="mx-3 flex items-center space-x-10">
+          {/* Recall Network Text */}
+          <span className="text-sm font-medium">RECALL.NETWORK</span>
+
+          <Button onClick={openConnectModal} className="bg-sky-700 px-6 py-5 text-white hover:bg-sky-600">
+            JOIN / SIGN IN
+          </Button>
+        </div>
+      )
+
+    }
+  }</ConnectButton.Custom>;
 };
+
